pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pytest-agentcontract"
version = "0.1.0"
description = "Deterministic CI tests for LLM agent trajectories — record once, replay offline, assert contracts"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Josh Pfizer" }]
keywords = [
    "pytest", "testing", "llm", "agents", "record-replay",
    "deterministic", "vcr", "langchain", "llamaindex", "openai",
    "anthropic", "ci", "regression"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Pytest",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pytest>=7.0",
    "pyyaml>=6.0",
    "jsonschema>=4.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.30"]
langchain = ["langchain-core>=0.2"]
llamaindex = ["llama-index-core>=0.10"]
all = [
    "openai>=1.0",
    "anthropic>=0.30",
    "langchain-core>=0.2",
    "llama-index-core>=0.10",
]
dev = [
    "ruff>=0.4",
    "mypy>=1.10",
    "pytest-cov>=5.0",
]

[project.entry-points.pytest11]
agentcontract = "agentcontract.plugin"

[project.scripts]
agentcontract = "agentcontract.cli:main"

[project.urls]
Homepage = "https://github.com/mikiships/pytest-agentcontract"
Repository = "https://github.com/mikiships/pytest-agentcontract"
Issues = "https://github.com/mikiships/pytest-agentcontract/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/agentcontract"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
